Fixed oauth cookie issue, and added ability to pass a connection around database related functions

+++ notilib/accounts.py
@@ -20,28 +20,15 @@ class AccountDataTuple(NamedTuple):
     blacklisted: int
 
 
-async def create_account(
+async def __create_account(
     discord_id: int,
-    account_id: int=None,
-    creation_timestamp: float=None,
-    permissions: list | str=None,
-    blacklisted: bool=False,
+    account_id: int,
+    creation_timestamp: float,
+    permissions: list | str,
+    blacklisted: bool,
+    conn: asyncpg.Pool
 ) -> bool:
-    """
-    Creates a new account for a user if ones doesn't already exist
-    :param discord_id: the discord id of the respective user
-    :param account_id: override the autoincrement account id
-    :param creation_timestamp: override the creation timestamp of the account
-    :param permissions: set the permissions for the user
-    :param blacklisted: set whether the accounts is blacklisted
-
-    :return: bool of whether or not the account was created
-    """
-    db = Database()
-    await db.connect()
-
-    # check if the account exists and return if it does
-    existing_account = await db.conn.fetchrow(
+    existing_account = await conn.fetchrow(
         'SELECT * FROM accounts WHERE discord_id = $1', discord_id)
 
     # return if account exists
@@ -73,25 +60,66 @@ async def create_account(
     column_names = ', '.join(data.keys())
     placeholders = create_query_placeholders(len(data.values()))
 
-    await db.conn.execute(
+    await conn.execute(
         f'INSERT INTO accounts ({column_names}) VALUES ({placeholders})',
         *tuple(data.values()))
 
     return True
 
 
-async def delete_account(discord_id: int, confirm: bool=False):
+async def create_account(
+    discord_id: int,
+    account_id: int=None,
+    creation_timestamp: float=None,
+    permissions: list | str=None,
+    blacklisted: bool=False,
+    conn: asyncpg.Pool=None
+) -> bool:
+    """
+    Creates a new account for a user if ones doesn't already exist
+    :param discord_id: the discord id of the respective user
+    :param account_id: override the autoincrement account id
+    :param creation_timestamp: override the creation timestamp of the account
+    :param permissions: set the permissions for the user
+    :param blacklisted: set whether the accounts is blacklisted
+    :param conn: an open database connection to execute on, if left as `None`,\
+        one will be acquired automatically
+
+    :return: bool of whether or not the account was created
+    """
+    if conn:  # use provided connection
+        return await __create_account(
+            discord_id, account_id, creation_timestamp, permissions, blacklisted, conn)
+
+    async with await Database().connect() as conn:  # otherwise, acquire new connection
+        # check if the account exists and return if it does
+        return await __create_account(
+            discord_id, account_id, creation_timestamp, permissions, blacklisted, conn)
+
+
+async def delete_account(
+    discord_id: int,
+    confirm: bool=False,
+    conn: asyncpg.Pool=None
+):
     """
     Entirely erase an account from existence
     :param discord_id: the discord id of the account to delete
     :param confirm: must `True` in order to delete the account
+    :param conn: an open database connection to execute on, if left as `None`,\
+        one will be acquired automatically
     """
+    # raise error if deletion is unconfirmed (making it harder to slip up)
     if confirm is False:
         raise ConfirmationError(
             'Param `confirm` must be `True` in order to delete an account!')
 
-    conn = await Database().connect()
-    await conn.execute('DELETE FROM accounts WHERE discord_id = $1', discord_id)
+    if conn:
+        await conn.execute('DELETE FROM accounts WHERE discord_id = $1', discord_id)
+        return
+
+    async with await Database().connect() as conn:
+        await conn.execute('DELETE FROM accounts WHERE discord_id = $1', discord_id)
 
 
 async def __select_account_data(discord_id: int, conn: asyncpg.Connection):
@@ -100,19 +128,12 @@ async def __select_account_data(discord_id: int, conn: asyncpg.Connection):
     return row
 
 
-async def get_account(
+async def __get_account(
     discord_id: int,
-    create: bool=False
+    create: bool,
+    conn: asyncpg.Pool
 ) -> AccountDataTuple | None:
-    """
-    Gets the account data for a user
-    :param discord_id: the discord id of the user to get the account data for
-    :param create: if an account should be created if one does not already exists
-    """
-    db = Database()
-    await db.connect()
-
-    account_data = await __select_account_data(discord_id, db.conn)
+    account_data = await __select_account_data(discord_id, conn)
 
     if account_data is None:
         # return no data if "create if not exists" is False
@@ -121,17 +142,57 @@ async def get_account(
 
         # create account and select account data
         await create_account(discord_id)
-        account_data = await __select_account_data(discord_id, db.conn)
+        account_data = await __select_account_data(discord_id, conn)
 
     # return account data tuple as a named tuple
     return AccountDataTuple(*account_data)
 
 
+async def get_account(
+    discord_id: int,
+    create: bool=False,
+    conn: asyncpg.Pool=None
+) -> AccountDataTuple | None:
+    """
+    Gets the account data for a user
+    :param discord_id: the discord id of the user to get the account data for
+    :param create: if an account should be created if one does not already exists
+    :param conn: an open database connection to execute on, if left as `None`,\
+        one will be acquired automatically
+    """
+    if conn:
+        return await __get_account(discord_id, create, conn)
+
+    async with await Database().connect() as conn:
+        return await __get_account(discord_id, create, conn)
+
+
+async def __set_account_blacklist(
+    discord_id: int,
+    blacklisted: bool,
+    create: bool,
+    conn: asyncpg.Pool
+):
+    # check if account exists
+    account_data = await __select_account_data(discord_id, conn)
+    if not account_data:
+        # create account with blacklist setting if "create if not exists" is true
+        if create:
+            await create_account(discord_id, blacklisted=blacklisted, conn=conn)
+        return
+
+    # update existing account blacklist value
+    await conn.execute(
+        'UPDATE accounts SET blacklisted = $1 WHERE discord_id = $2',
+        int(blacklisted), discord_id)
+
+
 async def set_account_blacklist(
     discord_id: int,
     blacklisted: bool=True,
-    create: bool=True
-):
+    create: bool=True,
+    conn: asyncpg.Pool=None
+) -> None:
     """
     Set the blacklist property of an account
     :param discord_id: the discord id of the user account
@@ -139,18 +200,9 @@ async def set_account_blacklist(
     :param blacklisted: whether to blacklist or unblacklist the user
     :param create: whether or not to create the account if it doesn't exist
     """
-    db = Database()
-    await db.connect()
-
-    # check if account exists
-    account_data = await __select_account_data(discord_id, db.conn)
-    if not account_data:
-        # create account with blacklist setting if "create if not exists" is true
-        if create:
-            await create_account(discord_id, blacklisted=blacklisted)
+    if conn:
+        await __set_account_blacklist(discord_id, blacklisted, create, conn)
         return
 
-    # update existing account blacklist value
-    await db.conn.execute(
-        'UPDATE accounts SET blacklisted = $1 WHERE discord_id = $2',
-        int(blacklisted), discord_id)
+    async with await Database().connect() as conn:
+        await __set_account_blacklist(discord_id, blacklisted, create, conn)
